Add files via upload
1. Are there any times when you can jump when you shouldn’t be able
to?
Yes. When the player character is in mid-jump the player is able to still jump further into the air the more they press the jump button. Most designers do not want to allow players to do this because then it not really functioning as a jump. 

2. Be able to answer why our player can jump while touching these locations.
Because our if statement in the update section checks if the player pressed space and our boolean “canJump” is true. Then we have two functions that check if the player is in contact with “Floor”, and if so makes “canJump” true and if not, makes it false.